Voxelization fix
"Exploding" collider was spawned for each voxel in grid! Luckily the
colliders' pool was only of size 25, so only 25 exploding colliders were
spawned each time.
Anyway this impacts on performance and was wrong.
Now a single collider is spawned into the grid center.
In addition, collider now can be resized to tune the explosion effect.
With this fix is not longer needed to random rotate every voxel to force
collision between them.