Blink script fix
Can not update materials while blinking because it would store the
current (wrong) materials as default ones.
There is a bug that needs to be solved though. If a color changing
material is updated while blinking (ie. player color goes from red to
green), when blinking ends red color is restored.